+++ src/components/Homepage.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Footer from './Footer'
+import Footer from './Footer/Footer'
 
 function Homepage() {
   return (
