--- src/index.css
@@ -1,149 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Archivo:wght@400;500;600;700&display=swap");
-
-*{
-    margin: 0;
-    padding: 0;
-}
-body{
-    min-height: 100vh;
-}
-
-.footer {
-   height: fit-content;
-   width: 100vw;
-   background-color: rgba(0, 0, 0, 0.5);
-   color: whitesmoke;
-   position: fixed;
-   bottom: 0;
-   line-height: 1.5;
-   text-align: center;
-}
-
-.main{
-    display: flex;
-   align-items: center;
-   justify-content: space-between;
-   padding: 2em;
-   
-}
-.footnote {
-    max-width: 500px;
-    margin: 0 auto;
-}
-.link_list {
-    text-decoration: none;
-    color: inherit;
-}
-.list {
-    list-style-type: none;
-    
-}
-@media (max-width: 768px) {
-    .footer{
-        position: fixed;
-        top: 50vh;
-        left: 0;
-        width: fit-content;
-        height: fit-content;
-        text-align: start;
-        padding: 0;
-    }
-    .news, .contacts, .year, .follow {
-        display: none;
-    }
-    
-}
-@media (max-width:768px){
-
-/* .filtershow {
-    position: absolute;
-    width: 262px;
-    height: 37px;
-    left: 30px;
-    top: 1246px;
-} */
-
-.filtercard{
-    position: relative;
-    width: 414px;
-    height: 896px;
-    top: -1788;
-    background: #FFFFFF;
-    
-} 
-.filtertop {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-.filterBtn{
-    position: absolute;
-    width: 29px;
-    height: 29px;
-    left: 355px;
-    top: 1249px;;
-}
-.filterheading {
-    position: absolute;
-    width: 262px;
-    height: 37px;
-    left: 30px;
-    top: 1246px;
-    font-family: Archivo;
-    font-style: normal;
-    font-weight: 700;
-    font-size: 18px;
-    line-height: 20x;
-
-
-}
-.filterheading_heavy{
-    width: 112px;
-    height: 20px;
-    left: 30px;
-    top: 1254px;
-    font-weight: 700;
-    color: #000000;
-}
-.filterheading_light{
-    width: 134px;
-    height: 20px;
-    left: 158px;
-    top: 1254px;
-    font-weight: 400;
-    color: #9B9B9B;
-}
-.filter_range {
-    position: absolute;
-    width: 414px;
-    height: 1484px;
-    left: 0px;
-    top: 88px;
-    background: #FFFFFF;
-}
-.filter_inner {
-    position: absolute;
-    width: 281px;
-    height: 1440px;
-    left: 36px;
-    top: 107px;
-}
-.filter_header {
-    display: flex;
-    justify-content: space-between;
-}
-input {
-    position: absolute;
-    width: 37px;
-    height: 37px;
-    border: 4px solid #000000;
-}
-.list{
-    font-family: 'Archivo';
-    font-style: normal;
-    font-weight: 400;
-    font-size: 28px;
-    line-height: 30px;
-    color: #1D1D1D;
-}
-}
