+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import './index.css'
+import './components/Footer/footer.css'
 import App from './App'
 
 ReactDOM.render(
